Move create arbitrary state method and console interface to own file
